hw1updated

+++ FVCalc.java
@@ -3,7 +3,7 @@ public class FVCalc {
 	public static void main(String[] args){
 		
         int currentValue = Integer.parseInt(args[0]);
-		int rate = Integer.parseInt(args[1]);
+		double rate = Double.parseDouble(args[1]);
 		int n = Integer.parseInt(args[2]);
 		double futureValue = currentValue * Math.pow(1.0 + (rate / 100.0), n);
         System.out.println("After " + n + " years, $" + currentValue + " saved at "
